refactor: split toast components to separate files (#41)
* refactor: remove unused import

* refactor: move title to separate file

* refactor: move close to separate file

* refactor: move root to separate file

* refactor: move viewport to separate file
