+++ site/src/features/toast/components/Title.tsx
@@ -1,0 +1,6 @@
+import { ToastTitle } from '@radix-ui/react-toast'
+import { styled } from '@config/theme'
+
+export const Title = styled(ToastTitle, {
+  margin: 'auto'
+})
